+++ blockfrost-api/src/Blockfrost/Types/Cardano/Pools.hs
@@ -44,6 +44,7 @@ data PoolInfo = PoolInfo
   , _poolInfoLiveStake      :: Lovelaces
   , _poolInfoLiveSize       :: Double
   , _poolInfoLiveSaturation :: Double
+  , _poolInfoLiveDelegators :: Double
   , _poolInfoActiveStake    :: Lovelaces
   , _poolInfoActiveSize     :: Double
   , _poolInfoDeclaredPledge :: Lovelaces -- ^ Stake pool certificate pledge
@@ -69,6 +70,7 @@ instance ToSample PoolInfo where
       , _poolInfoLiveStake = 6900000000
       , _poolInfoLiveSize = 0.42
       , _poolInfoLiveSaturation = 0.93
+      , _poolInfoLiveDelegators = 127
       , _poolInfoActiveStake = 4200000000
       , _poolInfoActiveSize = 0.43
       , _poolInfoDeclaredPledge = 5000000000