+++ blockfrost-api/src/Blockfrost/Types/Shared.hs
@@ -8,6 +8,7 @@ module Blockfrost.Types.Shared
   , module Blockfrost.Types.Shared.BlockHash
   , module Blockfrost.Types.Shared.BlockIndex
   , module Blockfrost.Types.Shared.CBOR
+  , module Blockfrost.Types.Shared.DatumHash
   , module Blockfrost.Types.Shared.Epoch
   , module Blockfrost.Types.Shared.Opts
   , module Blockfrost.Types.Shared.POSIXMillis
@@ -15,6 +16,7 @@ module Blockfrost.Types.Shared
   , module Blockfrost.Types.Shared.PoolId
   , module Blockfrost.Types.Shared.Quantity
   , module Blockfrost.Types.Shared.Slot
+  , module Blockfrost.Types.Shared.ScriptHash
   , module Blockfrost.Types.Shared.TxHash
   ) where
 
@@ -25,6 +27,7 @@ import Blockfrost.Types.Shared.AssetId
 import Blockfrost.Types.Shared.BlockHash
 import Blockfrost.Types.Shared.BlockIndex
 import Blockfrost.Types.Shared.CBOR
+import Blockfrost.Types.Shared.DatumHash
 import Blockfrost.Types.Shared.Epoch
 import Blockfrost.Types.Shared.Opts
 import Blockfrost.Types.Shared.POSIXMillis
@@ -32,4 +35,5 @@ import Blockfrost.Types.Shared.PolicyId
 import Blockfrost.Types.Shared.PoolId
 import Blockfrost.Types.Shared.Quantity
 import Blockfrost.Types.Shared.Slot
+import Blockfrost.Types.Shared.ScriptHash
 import Blockfrost.Types.Shared.TxHash