+++ blockfrost-api/test/Cardano/Addresses.hs
@@ -33,9 +33,9 @@ spec_sample = do
     Right addressDetailsExpected
 
   it "parses address utxos sample" $ do
-    eitherDecode addressUTXOsSample
+    eitherDecode addressUtxosSample
     `shouldBe`
-    Right addressUTXOsExpected
+    Right addressUtxosExpected
 
   it "parses address transactions sample" $ do
     eitherDecode addressTransactionsSample
@@ -56,7 +56,8 @@ addressInfoSample = [r|
     }
   ],
   "stake_address": "stake1ux3g2c9dx2nhhehyrezyxpkstartcqmu9hk63qgfkccw5rqttygt7",
-  "type": "shelley"
+  "type": "shelley",
+  "script": false
 }
 |]
 
@@ -73,6 +74,7 @@ addressInfoExpected =
       ]
     , _addressInfoStakeAddress = pure "stake1ux3g2c9dx2nhhehyrezyxpkstartcqmu9hk63qgfkccw5rqttygt7"
     , _addressInfoType = Shelley
+    , _addressInfoScript = False
     }
 
 addressDetailsSample = [r|
@@ -119,7 +121,7 @@ addressDetailsExpected =
             ]
 
 
-addressUTXOsSample = [r|
+addressUtxosSample = [r|
 [
     {
         "tx_hash": "39a7a284c2a0948189dc45dec670211cd4d72f7b66c5726c08d9b3df11e44d58",
@@ -131,7 +133,8 @@ addressUTXOsSample = [r|
                 "quantity": "42000000"
             }
         ],
-        "block": "7eb8e27d18686c7db9a18f8bbcfe34e3fed6e047afaa2d969904d15e934847e6"
+        "block": "7eb8e27d18686c7db9a18f8bbcfe34e3fed6e047afaa2d969904d15e934847e6",
+        "data_hash": "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710"
     },
     {
         "tx_hash": "4c4e67bafa15e742c13c592b65c8f74c769cd7d9af04c848099672d1ba391b49",
@@ -143,7 +146,8 @@ addressUTXOsSample = [r|
                 "quantity": "729235000"
             }
         ],
-        "block": "953f1b80eb7c11a7ffcd67cbd4fde66e824a451aca5a4065725e5174b81685b7"
+        "block": "953f1b80eb7c11a7ffcd67cbd4fde66e824a451aca5a4065725e5174b81685b7",
+        "data_hash": null
     },
     {
         "tx_hash": "768c63e27a1c816a83dc7b07e78af673b2400de8849ea7e7b734ae1333d100d2",
@@ -159,28 +163,31 @@ addressUTXOsSample = [r|
                 "quantity": "12"
             }
         ],
-        "block": "5c571f83fe6c784d3fbc223792627ccf0eea96773100f9aedecf8b1eda4544d7"
+        "block": "5c571f83fe6c784d3fbc223792627ccf0eea96773100f9aedecf8b1eda4544d7",
+        "data_hash": null
     }
 ]
 |]
 
-addressUTXOsExpected =
-  [ AddressUTXO
-    { _addressUTXOTxHash = "39a7a284c2a0948189dc45dec670211cd4d72f7b66c5726c08d9b3df11e44d58"
-    , _addressUTXOOutputIndex = 0
-    , _addressUTXOAmount = [ AdaAmount 42000000 ]
-    , _addressUTXOBlock = "7eb8e27d18686c7db9a18f8bbcfe34e3fed6e047afaa2d969904d15e934847e6"
+addressUtxosExpected =
+  [ AddressUtxo
+    { _addressUtxoTxHash = "39a7a284c2a0948189dc45dec670211cd4d72f7b66c5726c08d9b3df11e44d58"
+    , _addressUtxoOutputIndex = 0
+    , _addressUtxoAmount = [ AdaAmount 42000000 ]
+    , _addressUtxoBlock = "7eb8e27d18686c7db9a18f8bbcfe34e3fed6e047afaa2d969904d15e934847e6"
+    , _addressUtxoDataHash = Just "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710"
     }
-  , AddressUTXO
-    { _addressUTXOTxHash = "4c4e67bafa15e742c13c592b65c8f74c769cd7d9af04c848099672d1ba391b49"
-    , _addressUTXOOutputIndex = 0
-    , _addressUTXOAmount = [ AdaAmount 729235000 ]
-    , _addressUTXOBlock = "953f1b80eb7c11a7ffcd67cbd4fde66e824a451aca5a4065725e5174b81685b7"
+  , AddressUtxo
+    { _addressUtxoTxHash = "4c4e67bafa15e742c13c592b65c8f74c769cd7d9af04c848099672d1ba391b49"
+    , _addressUtxoOutputIndex = 0
+    , _addressUtxoAmount = [ AdaAmount 729235000 ]
+    , _addressUtxoBlock = "953f1b80eb7c11a7ffcd67cbd4fde66e824a451aca5a4065725e5174b81685b7"
+    , _addressUtxoDataHash = Nothing
     }
-  , AddressUTXO
-    { _addressUTXOTxHash = "768c63e27a1c816a83dc7b07e78af673b2400de8849ea7e7b734ae1333d100d2"
-    , _addressUTXOOutputIndex = 1
-    , _addressUTXOAmount =
+  , AddressUtxo
+    { _addressUtxoTxHash = "768c63e27a1c816a83dc7b07e78af673b2400de8849ea7e7b734ae1333d100d2"
+    , _addressUtxoOutputIndex = 1
+    , _addressUtxoAmount =
         [ AdaAmount 42000000
         , AssetAmount
             $ Money.mkSomeDiscrete
@@ -188,7 +195,8 @@ addressUTXOsExpected =
                  unitScale
                  12
         ]
-    , _addressUTXOBlock = "5c571f83fe6c784d3fbc223792627ccf0eea96773100f9aedecf8b1eda4544d7"
+    , _addressUtxoBlock = "5c571f83fe6c784d3fbc223792627ccf0eea96773100f9aedecf8b1eda4544d7"
+    , _addressUtxoDataHash = Nothing
     }
   ]